Force marketcap to be an integer to prevent scientific notation

+++ src/rpc.cpp
@@ -1170,7 +1170,7 @@ void RPC::refreshPrice() {
                     ui->volumeLocal->setText( QString::number((double) vol / (double) price) + " HUSH");
 
                 qDebug() << "Mcap = " << (double) mcap;
-                ui->marketcap->setText(  QString::number( (double) mcap) + " " + QString::fromStdString(ticker) );
+                ui->marketcap->setText(  QString::number( (unsigned int) mcap) + " " + QString::fromStdString(ticker) );
                 ui->marketcapBTC->setText( QString::number((double) btcmcap) + " BTC" );
                 //ui->marketcapLocal->setText( QString::number((double) mcap * (double) price) + " " + QString::fromStdString(ticker) );
